// UI/ang-persona/src/app/Models/Persona.ts
export class PersonaVM {
    nombre: string | undefined;
    apellido: string | undefined;
    edad: number | undefined;
    constructor() {
    }
}

export class Personas {
    id: string | undefined;
    nombre: string | undefined;
    apellido: string | undefined;
    edad: number | undefined;
    constructor() {
    }
}